Foreign location names mapped to standard name in HW inv tables
