+++ dai_network_listener/src/test/groovy/com/intel/dai/network_listener/BenchmarkingSystemActionsSpec.groovy
@@ -83,10 +83,10 @@ class BenchmarkingSystemActionsSpec extends Specification {
         expect: true
     }
 
-    def "upsertHWInventory"() {
-        underTest_.upsertHWInventory("")
-        expect: true
-    }
+//    def "upsertHWInventory"() {
+//        underTest_.upsertHWInventory("")
+//        expect: true
+//    }
 
     def "Close"() {
         underTest_.close()
