+++ docker-deploy/etc/HWInventory.json
